# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SalonSeal API - No-show deposit system for Nairobi salons
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings (filterable by salonId, status, staffId)
      parameters:
        - name: salonId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: date
          in: query
          required: false
          schema:
            type: string
        - name: staffId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking (client-facing)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created booking(s)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateBookingResponse"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
  /bookings/{id}/status:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status (arrived, no-show, completed, cancelled)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingStatusBody"
      responses:
        "200":
          description: Updated booking with auto-blacklist info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateBookingStatusResponse"
  /bookings/{id}/simulate-payment:
    post:
      operationId: simulateMpesaPayment
      tags: [bookings]
      summary: Simulate an M-Pesa deposit payment (dev/demo use)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulatePaymentResponse"
        "400":
          description: Already paid or invalid state
        "404":
          description: Not found
  /bookings/{id}/cancel:
    post:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking (checks cancellation window, triggers refund if eligible)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelBookingBody"
      responses:
        "200":
          description: Cancellation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancellationResult"
  /bookings/{id}/send-deposit-nudge:
    post:
      operationId: sendDepositNudge
      tags: [bookings]
      summary: Send a WhatsApp M-Pesa payment request to client for unpaid deposit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Nudge sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositNudgeResponse"
        "400":
          description: Deposit already paid or booking in invalid state
        "404":
          description: Not found
  /bookings/{id}/staff:
    patch:
      operationId: assignBookingStaff
      tags: [bookings]
      summary: Assign or reassign a staff member to a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignBookingStaffBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
  /bookings/{id}/reschedule:
    patch:
      operationId: rescheduleBooking
      tags: [bookings]
      summary: Reschedule a booking to a new date/time
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleBookingBody"
      responses:
        "200":
          description: Rescheduled booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found

  # ── Waitlist ─────────────────────────────────────────────────────────────────
...
  /salons/{salonId}/staff/{staffId}/busy-slots:
    get:
      operationId: getStaffBusySlots
      tags: [staff]
      summary: Get booked appointment times for a staff member on a given date
      parameters:
        - name: salonId
          in: path
          required: true
          schema:
            type: integer
        - name: staffId
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of busy time slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffBusySlotsResponse"
...
components:
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        salonId:
          type: integer
        serviceId:
          type: integer
        staffId:
          type: ["integer", "null"]
        clientId:
          type: integer
        clientName:
          type: string
        clientPhone:
          type: string
        appointmentAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, confirmed, arrived, completed, cancelled, no_show]
        depositAmount:
          type: integer
        depositPaid:
          type: boolean
        depositWaived:
          type: boolean
        mpesaRef:
          type: ["string", "null"]
        refundEligible:
          type: boolean
        notes:
          type: ["string", "null"]
        serviceName:
          type: string
        staffName:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
        recurringGroupId:
          type: ["integer", "null"]
        recurrenceRule:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - salonId
        - serviceId
        - clientId
        - clientName
        - clientPhone
        - appointmentAt
        - status
        - depositAmount
        - depositPaid
        - depositWaived
        - refundEligible
        - serviceName
        - createdAt
    CreateBookingBody:
      type: object
      properties:
        salonId:
          type: integer
        serviceId:
          type: integer
        staffId:
          type: ["integer", "null"]
        clientName:
          type: string
        clientPhone:
          type: string
        appointmentAt:
          type: string
          format: date-time
        notes:
          type: ["string", "null"]
        depositWaived:
          type: boolean
        recurrenceRule:
          type: string
          enum: [weekly, biweekly, monthly]
        recurrenceCount:
          type: integer
      required:
        - salonId
        - serviceId
        - clientName
        - clientPhone
        - appointmentAt
    CreateBookingResponse:
      type: object
      properties:
        booking:
          $ref: "#/components/schemas/Booking"
        additionalBookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
      required:
        - booking
        - additionalBookings
    AssignBookingStaffBody:
      type: object
      properties:
        staffId:
          type: ["integer", "null"]
      required:
        - staffId
    UpdateBookingStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [arrived, completed, no_show, cancelled]
      required:
        - status
    UpdateBookingStatusResponse:
      type: object
      properties:
        booking:
          $ref: "#/components/schemas/Booking"
        autoBlacklisted:
          type: boolean
        autoBlacklistedName:
          type: ["string", "null"]
      required:
        - booking
        - autoBlacklisted
        - autoBlacklistedName
    CancelBookingBody:
      type: object
      properties:
        reason:
          type: ["string", "null"]
      required: []
    CancellationResult:
      type: object
      properties:
        booking:
          $ref: "#/components/schemas/Booking"
        refundEligible:
          type: boolean
        refundAmount:
          type: integer
        message:
          type: string
        waitlistedClients:
          type: array
          items:
            $ref: "#/components/schemas/WaitlistEntry"
        autoPromoted:
          type: boolean
        autoPromotedClientName:
          type: ["string", "null"]
        autoPromotedPhone:
          type: ["string", "null"]
        autoPromotedMessage:
          type: ["string", "null"]
      required:
        - booking
        - refundEligible
        - refundAmount
        - message
        - waitlistedClients
        - autoPromoted
...
    RescheduleBookingBody:
      type: object
      properties:
        appointmentAt:
          type: string
          format: date-time
      required:
        - appointmentAt
    StaffBusySlot:
      type: object
      properties:
        appointmentAt:
          type: string
          format: date-time
        durationMinutes:
          type: integer
      required:
        - appointmentAt
        - durationMinutes
    StaffBusySlotsResponse:
      type: object
      properties:
        busySlots:
          type: array
          items:
            $ref: "#/components/schemas/StaffBusySlot"
      required:
        - busySlots
...
    WaitlistEntry:
      type: object
      properties:
        id:
          type: integer
        salonId:
          type: integer
        staffId:
          type: ["integer", "null"]
        serviceId:
          type: ["integer", "null"]
        appointmentAt:
          type: string
          format: date-time
        clientName:
          type: string
        clientPhone:
          type: string
        serviceName:
          type: string
        staffName:
          type: ["string", "null"]
        notified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - salonId
        - appointmentAt
        - clientName
        - clientPhone
        - serviceName
        - notified
        - createdAt
...
    SimulatePaymentResponse:
      type: object
      properties:
        booking:
          $ref: "#/components/schemas/Booking"
        mpesaRef:
          type: string
        message:
          type: string
      required:
        - booking
        - mpesaRef
        - message
    DepositNudgeResponse:
      type: object
      properties:
        message:
          type: string
        waMessageId:
          type: string
      required:
        - message
        - waMessageId
...